requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
email-validator==2.2.0
